% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d10misc-package.R
\docType{package}
\name{d10misc-package}
\alias{d10misc}
\alias{d10misc-package}
\title{d10misc: Miscellaneous tools}
\description{
Miscellaneous tools.
}
\author{
\strong{Maintainer}: Diego Diez \email{diego10ruiz@gmail.com}

}
\keyword{internal}
