% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_heatmap}
\alias{plot_heatmap}
\alias{plot_heatmap.matrix}
\title{plot_heatmap}
\usage{
plot_heatmap(x, ...)

\method{plot_heatmap}{matrix}(x, row.cluster = TRUE,
  col.cluster = FALSE, scale = TRUE, row.name = "row",
  col.name = "col", value.name = "value", ...)
}
\arguments{
\item{x}{object to plot.}

\item{...}{arguments passed down to methods.}

\item{row.cluster}{logical; whether to cluster rows.}

\item{col.cluster}{logical; whether to cluster columns}

\item{scale}{logical; whether to scale rows.}

\item{row.name}{name for the row variable.}

\item{col.name}{name for the col variable.}

\item{value.name}{name for the value variable.}
}
\description{
plot_heatmap
}
