% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{remove_axis_text}
\alias{remove_axis_text}
\title{remove_axis_text}
\usage{
remove_axis_text(type = c("x", "y"))
}
\arguments{
\item{type}{the axes to remove the text from.}
}
\description{
remove_axis_text
}
