% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{remove_axis_ticks}
\alias{remove_axis_ticks}
\title{remove_axis_ticks}
\usage{
remove_axis_ticks(type = c("x", "y"))
}
\arguments{
\item{type}{the axes to remove the ticks from.}
}
\description{
remove_axis_ticks
}
