% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{rotate_xlab}
\alias{rotate_xlab}
\title{rotate_xlab}
\usage{
rotate_xlab(angle = 90, hjust = 1, vjust = 0.5, ...)
}
\arguments{
\item{angle}{angle to rotate x labels.}

\item{hjust}{horizontal justification.}

\item{vjust}{vertical justification.}

\item{...}{arguments passed down to element_text().}
}
\description{
Sets element_text() values for axis.text.x so that horizontal x-labels
are perpendicular and nicely aligned with the x-axis.
}
