% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_kuro}
\alias{theme_kuro}
\title{theme_kuro}
\usage{
theme_kuro(base_size = 16, ...)
}
\arguments{
\item{base_size}{base size for text.}

\item{...}{arguments passed down to theme_dark().}
}
\description{
A tweaked ggplot2::theme_dark().
}
