% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_base}
\alias{theme_base}
\title{theme_base}
\usage{
theme_base(base_size = 12, ...)
}
\arguments{
\item{base_size}{base size for text.}

\item{...}{arguments passed down to theme_dark().}
}
\description{
A tweaked ggthemes::theme_base().
}
